+++ pysatSpaceWeather/instruments/methods/__init__.py
@@ -2,5 +2,7 @@ from pysatSpaceWeather.instruments.methods import ace  # noqa F401
 from pysatSpaceWeather.instruments.methods import dst  # noqa F401
 from pysatSpaceWeather.instruments.methods import f107  # noqa F401
 from pysatSpaceWeather.instruments.methods import general  # noqa F401
+from pysatSpaceWeather.instruments.methods import gfz  # noqa F401
 from pysatSpaceWeather.instruments.methods import kp_ap  # noqa F401
 from pysatSpaceWeather.instruments.methods import lisird # noqa F401
+from pysatSpaceWeather.instruments.methods import swpc  # noqa F401