+++ pysatSpaceWeather/instruments/sw_kp.py
@@ -11,18 +11,23 @@ tag
     - '' Deprecated, mixed definitive and nowcast Kp data from GFZ
     - 'def' Definitive Kp data from GFZ
     - 'now' Nowcast Kp data from GFZ
-    - 'forecast' Grab forecast data from SWPC (next 3 days)
-    - 'recent' Grab last 30 days of Kp data from SWPC
+    - 'prediction' Predictions from SWPC for the next 3 days
+    - 'forecast' Forecast data from SWPC (next 3 days)
+    - 'recent' The last 30 days of Kp data from SWPC
 inst_id
     ''
 
 Note
 ----
-Downloads data from ftp.gfz-potsdam.de or SWPC.
+Downloads data from ftp.gfz-potsdam.de or SWPC. These files also contain ap
+data (with the GFZ data additionally containing Cp data), and so the additional
+data files will be saved to the appropriate data directories to avoid multiple
+downloads.
 
-Standard Kp files are stored by the first day of each month. When downloading
-use kp.download(start, stop, freq='MS') to only download days that could
-possibly have data.  'MS' gives a monthly start frequency.
+The historic definitive and nowcast Kp files are stored in yearly files, with
+the current year being updated remotely on a regular basis.  If you are using
+historic data for the current year, we recommend re-downloading it before
+performing your data processing.
 
 The forecast data is stored by generation date, where each file contains the
 forecast for the next three days. Forecast data downloads are only supported
@@ -46,10 +51,10 @@ Examples
 
 Warnings
 --------
-The 'forecast' and 'recent' tags load Kp data for a specific period of time.
-Loading multiple files, loading multiple days, the data padding feature, and
-multi_file_day feature available from the pyast.Instrument object is not
-appropriate for these tags data.
+The 'forecast', 'prediction', and 'recent' tags load Kp data for a specific
+period of time. Loading multiple files, loading multiple days, the data padding
+feature, and multi_file_day feature available from the pyast.Instrument object
+is not appropriate for these tags data.
 
 This material is based upon work supported by the
 National Science Foundation under Grant Number 1259508.
@@ -62,15 +67,12 @@ of the National Science Foundation.
 
 import datetime as dt
 import numpy as np
-import os
 import pandas as pds
-import requests
 import warnings
 
 import pysat
 
-from pysatSpaceWeather.instruments.methods import general
-from pysatSpaceWeather.instruments.methods import kp_ap
+from pysatSpaceWeather.instruments import methods
 
 # ----------------------------------------------------------------------------
 # Instrument attributes
@@ -80,6 +82,7 @@ name = 'kp'
 tags = {'': 'Deprecated, mixed definitive and nowcast Kp data from GFZ',
         'def': 'Definitive Kp data from GFZ',
         'now': 'Nowcast Kp data from GFZ',
+        'prediction': 'SWPC Predictions for the next three days',
         'forecast': 'SWPC Forecast data next (3 days)',
         'recent': 'SWPC provided Kp for past 30 days'}
 inst_ids = {'': list(tags.keys())}
@@ -95,6 +98,7 @@ today = dt.datetime(now.year, now.month, now.day)
 _test_dates = {'': {'def': dt.datetime(2009, 1, 1),
                     'now': dt.datetime(2020, 1, 1),
                     'forecast': today + dt.timedelta(days=1),
+                    'prediction': today + dt.timedelta(days=1),
                     'recent': today}}
 
 # Other tags assumed to be True
@@ -103,14 +107,14 @@ _test_download_ci = {'': {'': False}}
 # ----------------------------------------------------------------------------
 # Instrument methods
 
-preprocess = general.preprocess
+preprocess = methods.general.preprocess
 
 
 def init(self):
     """Initialize the Instrument object with instrument specific values."""
 
-    self.acknowledgements = kp_ap.acknowledgements(self.name, self.tag)
-    self.references = kp_ap.references(self.name, self.tag)
+    self.acknowledgements = methods.kp_ap.acknowledgements(self.name, self.tag)
+    self.references = methods.kp_ap.references(self.name, self.tag)
     pysat.logger.info(self.acknowledgements)
 
     if self.tag in ["def", "now"]:
@@ -150,7 +154,7 @@ def load(fnames, tag='', inst_id=''):
 
     Returns
     -------
-    data : pandas.DataFrame
+    result : pandas.DataFrame
         Object containing satellite data
     meta : pysat.Meta
         Object containing metadata such as column names and units
@@ -274,42 +278,10 @@ def load(fnames, tag='', inst_id=''):
         # files, and we need to return data daily.  The daily date is
         # attached to filename.  Parse off the last date, load month of data,
         # and downselect to the desired day
-        data = pds.DataFrame()
-
-        unique_fnames = dict()
-        for filename in fnames:
-            fname = filename[0:-11]
-            fdate = dt.datetime.strptime(filename[-10:], '%Y-%m-%d')
-            if fname not in unique_fnames.keys():
-                unique_fnames[fname] = [fdate]
-            else:
-                unique_fnames[fname].append(fdate)
-
-        # Load the desired filenames
-        all_data = []
-        for fname in unique_fnames.keys():
-            # The daily date is attached to the filename.  Parse off the last
-            # date, load the year of data, downselect to the desired day
-            fdate = min(unique_fnames[fname])
-            temp = pds.read_csv(fname, index_col=0, parse_dates=True)
-
-            if temp.empty:
-                pysat.logger.warn('Empty file: {:}'.format(fname))
-                continue
-
-            # Select the desired times and add to data list
-            all_data.append(pds.DataFrame(temp[fdate:max(unique_fnames[fname])
-                                               + dt.timedelta(seconds=86399)]))
-
-        # Combine data together
-        if len(all_data) > 0:
-            result = pds.concat(all_data, axis=0, sort=True)
-        else:
-            result = pds.DataFrame()
-
+        result = methods.gfz.load_def_now(fnames)
         fill_val = np.nan
     else:
-        # Load the forecast or recent data
+        # Load the prediction, forecast or recent data
         all_data = []
         for fname in fnames:
             result = pds.read_csv(fname, index_col=0, parse_dates=True)
@@ -319,50 +291,15 @@ def load(fnames, tag='', inst_id=''):
         fill_val = -1
 
     # Initalize the meta data
-    if tag in ['', 'forecast', 'recent']:
+    if tag in ['', 'forecast', 'recent', 'prediction']:
         for kk in result.keys():
-            kp_ap.initialize_kp_metadata(meta, kk, fill_val)
+            methods.kp_ap.initialize_kp_metadata(meta, kk, fill_val)
     else:
         for kk in result.keys():
             if kk.find('Kp') >= 0:
-                kp_ap.initialize_kp_metadata(meta, kk, fill_val)
-            elif kk.lower().find('ap') >= 0:
-                kp_ap.initialize_ap_metadata(meta, kk, fill_val)
-
-        meta['Bartels_solar_rotation_num'] = {
-            meta.labels.units: '',
-            meta.labels.name: 'Bartels solar rotation number',
-            meta.labels.desc: ''.join(['A sequence of 27-day intervals counted',
-                                       ' from February 8, 1832']),
-            meta.labels.min_val: 1,
-            meta.labels.max_val: np.inf,
-            meta.labels.fill_val: -1}
-        meta['day_within_Bartels_rotation'] = {
-            meta.labels.units: 'days',
-            meta.labels.name: 'Bartels solar rotation number',
-            meta.labels.desc: ''.join(['Number of day within the Bartels solar',
-                                       ' rotation']),
-            meta.labels.min_val: 1,
-            meta.labels.max_val: 27,
-            meta.labels.fill_val: -1}
-        meta['Cp'] = {
-            meta.labels.units: '',
-            meta.labels.name: 'Cp index',
-            meta.labels.desc: ''.join(['the daily planetary character figure, ',
-                                       'a qualitative estimate of the overall ',
-                                       'level of geomagnetic activity for ',
-                                       'this day determined from the sum of ',
-                                       'the eight ap amplitudes']),
-            meta.labels.min_val: 0.0,
-            meta.labels.max_val: 2.5,
-            meta.labels.fill_val: np.nan}
-        meta['C9'] = {
-            meta.labels.units: '',
-            meta.labels.name: 'C9 index',
-            meta.labels.desc: ''.join(['the contracted scale for Cp']),
-            meta.labels.min_val: 0,
-            meta.labels.max_val: 9,
-            meta.labels.fill_val: -1}
+                methods.kp_ap.initialize_kp_metadata(meta, kk, fill_val)
+            elif kk.find('Bartels') >= 0:
+                methods.kp_ap.initialize_bartel_metadata(meta, kk)
 
     return result, meta
 
@@ -420,29 +357,11 @@ def list_files(tag='', inst_id='', data_path='', format_str=None):
             files = files.asfreq('D', 'pad')
             files = files + '_' + files.index.strftime('%Y-%m-%d')
     elif tag in ['def', 'now']:
-        if format_str is None:
-            format_str = ''.join(['Kp_{:s}'.format(tag), '{year:04d}.txt'])
-
-        # Files are stored by year, going to add a date to the yearly
-        # filename for each month and day of month.  The load routine will load
-        # the year and use the append date to select out approriate data.
-        files = pysat.Files.from_os(data_path=data_path, format_str=format_str)
-        if not files.empty:
-            files.loc[files.index[-1] + pds.DateOffset(years=1)
-                      - pds.DateOffset(days=1)] = files.iloc[-1]
-            files = files.asfreq('D', 'pad')
-            files = files + '_' + files.index.strftime('%Y-%m-%d')
+        files = methods.gfz.kp_ap_cp_list_files(name, tag, inst_id, data_path,
+                                                format_str=format_str)
     else:
-        if format_str is None:
-            format_str = '_'.join(['kp', tag,
-                                   '{year:04d}-{month:02d}-{day:02d}.txt'])
-        files = pysat.Files.from_os(data_path=data_path, format_str=format_str)
-
-        # Pad list of files data to include most recent file under tomorrow
-        if not files.empty:
-            pds_offset = pds.DateOffset(days=1)
-            files.loc[files.index[-1] + pds_offset] = files.values[-1]
-            files.loc[files.index[-1] + pds_offset] = files.values[-1]
+        files = methods.swpc.list_files(name, tag, inst_id, data_path,
+                                        format_str=format_str)
 
     return files
 
@@ -484,191 +403,14 @@ def download(date_array, tag, inst_id, data_path):
                                "supported by GFZ."]),
                       DeprecationWarning, stacklevel=2)
     elif tag in ['def', 'now']:
-        # Set the page for the definitive or nowcast Kp
-        burl = ''.join(['https://datapub.gfz-potsdam.de/download/10.5880.Kp.',
-                        '0001/Kp_', 'nowcast' if tag == 'now' else 'definitive',
-                        '/'])
-        data_cols = ['Bartels_solar_rotation_num',
-                     'day_within_Bartels_rotation', 'Kp', 'daily_Kp_sum', 'ap',
-                     'daily_Ap', 'Cp', 'C9']
-        hours = np.arange(0, 24, 3)
-        kp_translate = {'0': 0.0, '3': 1.0 / 3.0, '7': 2.0 / 3.0}
-        dnames = list()
-
-        for dl_date in date_array:
-            fname = 'Kp_{:s}{:04d}.wdc'.format(tag, dl_date.year)
-            if fname not in dnames:
-                pysat.logger.info(' '.join(('Downloading file for',
-                                            dl_date.strftime('%Y'))))
-                furl = ''.join([burl, fname])
-                req = requests.get(furl)
-
-                if req.ok:
-                    # Split the file text into lines
-                    lines = req.text.split('\n')[:-1]
-
-                    # Remove the header
-                    while lines[0].find('#') == 0:
-                        lines.pop(0)
-
-                    # Process the data lines
-                    ddict = {dkey: list() for dkey in data_cols}
-                    times = list()
-                    for line in lines:
-                        ldate = dt.datetime.strptime(' '.join([
-                            "{:02d}".format(int(dd)) for dd in
-                            [line[:2], line[2:4], line[4:6]]]), "%y %m %d")
-                        bsr_num = np.int64(line[6:10])
-                        bsr_day = np.int64(line[10:12])
-                        if line[28:30] == '  ':
-                            kp_ones = 0.0
-                        else:
-                            kp_ones = np.float64(line[28:30])
-                        sum_kp = kp_ones + kp_translate[line[30]]
-                        daily_ap = np.int64(line[55:58])
-                        cp = np.float64(line[58:61])
-                        c9 = np.int64(line[61])
-
-                        for i, hour in enumerate(hours):
-                            # Set the time for this hour and day
-                            times.append(ldate + dt.timedelta(hours=int(hour)))
-
-                            # Set the daily values for this hour
-                            ddict['Bartels_solar_rotation_num'].append(bsr_num)
-                            ddict['day_within_Bartels_rotation'].append(bsr_day)
-                            ddict['daily_Kp_sum'].append(sum_kp)
-                            ddict['daily_Ap'].append(daily_ap)
-                            ddict['Cp'].append(cp)
-                            ddict['C9'].append(c9)
-
-                            # Get the hourly-specific values
-                            ikp = i * 2
-                            kp_ones = line[12 + ikp]
-                            if kp_ones == ' ':
-                                kp_ones = 0.0
-                            ddict['Kp'].append(np.float64(kp_ones)
-                                               + kp_translate[line[13 + ikp]])
-                            iap = i * 3
-                            ddict['ap'].append(np.int64(
-                                line[31 + iap:34 + iap]))
-
-                    # Put data into nicer DataFrame
-                    data = pds.DataFrame(ddict, index=times, columns=data_cols)
-
-                    # Write out as a CSV file
-                    saved_fname = os.path.join(data_path, fname).replace(
-                        '.wdc', '.txt')
-                    data.to_csv(saved_fname, header=True)
-
-                    # Record the filename so we don't download it twice
-                    dnames.append(fname)
-                else:
-                    pysat.logger.info("".join(["Unable to download data for ",
-                                               dl_date.strftime("%d %b %Y"),
-                                               ", date may be out of range ",
-                                               "for the database."]))
-
+        methods.gfz.kp_ap_cp_download(platform, name, date_array, tag, inst_id,
+                                      data_path)
     elif tag == 'forecast':
-        pysat.logger.info(' '.join(('This routine can only download the ',
-                                    'current forecast, not archived ',
-                                    'forecasts')))
-
-        # Download webpage
-        furl = 'https://services.swpc.noaa.gov/text/3-day-geomag-forecast.txt'
-        req = requests.get(furl)
-
-        # Parse text to get the date the prediction was generated
-        date_str = req.text.split(':Issued: ')[-1].split(' UTC')[0]
-        dl_date = dt.datetime.strptime(date_str, '%Y %b %d %H%M')
-
-        # Data is the forecast value for the next three days
-        raw_data = req.text.split('NOAA Kp index forecast ')[-1]
-
-        # Get date of the forecasts
-        date_str = raw_data[0:6] + ' ' + str(dl_date.year)
-        forecast_date = dt.datetime.strptime(date_str, '%d %b %Y')
-
-        # Strings we will use to parse the downloaded text
-        lines = ['00-03UT', '03-06UT', '06-09UT', '09-12UT', '12-15UT',
-                 '15-18UT', '18-21UT', '21-00UT']
-
-        # Storage for daily forecasts.
-        # Get values for each day, then combine together
-        day1 = []
-        day2 = []
-        day3 = []
-        for line in lines:
-            raw = raw_data.split(line)[-1].split('\n')[0]
-            cols = raw.split()
-            day1.append(np.float64(cols[-3]))
-            day2.append(np.float64(cols[-2]))
-            day3.append(np.float64(cols[-1]))
-
-        times = pds.date_range(forecast_date, periods=24, freq='3H')
-        day = []
-        for dd in [day1, day2, day3]:
-            day.extend(dd)
-
-        # Put data into nicer DataFrame
-        data = pds.DataFrame(day, index=times, columns=['Kp'])
-
-        # Write out as a file
-        data_file = 'kp_forecast_{:s}.txt'.format(dl_date.strftime('%Y-%m-%d'))
-        data.to_csv(os.path.join(data_path, data_file), header=True)
-
+        methods.swpc.geomag_forecast_download(name, date_array, data_path)
     elif tag == 'recent':
-        pysat.logger.info(' '.join(('This routine can only download the ',
-                                    'current webpage, not archived forecasts')))
-
-        # Download webpage
-        rurl = ''.join(('https://services.swpc.noaa.gov/text/',
-                        'daily-geomagnetic-indices.txt'))
-        req = requests.get(rurl)
-
-        # Parse text to get the date the prediction was generated
-        date_str = req.text.split(':Issued: ')[-1].split('\n')[0]
-        dl_date = dt.datetime.strptime(date_str, '%H%M UT %d %b %Y')
-
-        # Data is the forecast value for the next three days
-        raw_data = req.text.split('#  Date ')[-1]
-
-        # Keep only the middle bits that matter
-        raw_data = raw_data.split('\n')[1:-1]
-
-        # Hold times from the file
-        kp_time = []
-
-        # Holds Kp value for each station
-        sub_kps = [[], [], []]
-
-        # Iterate through file lines and parse out the info we want
-        for line in raw_data:
-            kp_time.append(dt.datetime.strptime(line[0:10], '%Y %m %d'))
-
-            # Pick out Kp values for each of the three columns. The columns
-            # used to all have integer values, but now some have floats.
-            sub_lines = [line[17:33], line[40:56], line[63:]]
-            for i, sub_line in enumerate(sub_lines):
-                split_sub = sub_line.split()
-                for ihr in np.arange(8):
-                    if sub_line.find('.') < 0:
-                        # These are integer values
-                        sub_kps[i].append(
-                            np.int64(sub_line[(ihr * 2):((ihr + 1) * 2)]))
-                    else:
-                        # These are float values
-                        sub_kps[i].append(np.float64(split_sub[ihr]))
-
-        # Create times on 3 hour cadence
-        times = pds.date_range(kp_time[0], periods=(8 * 30), freq='3H')
-
-        # Put into DataFrame
-        data = pds.DataFrame({'mid_lat_Kp': sub_kps[0],
-                              'high_lat_Kp': sub_kps[1],
-                              'Kp': sub_kps[2]}, index=times)
-
-        # Write out as a file
-        data_file = 'kp_recent_{:s}.txt'.format(dl_date.strftime('%Y-%m-%d'))
-        data.to_csv(os.path.join(data_path, data_file), header=True)
+        methods.swpc.kp_ap_recent_download(name, date_array, data_path)
+    elif tag == 'prediction':
+        methods.swpc.solar_geomag_predictions_download(name, date_array,
+                                                       data_path)
 
     return