+++ CHANGELOG.md
@@ -3,6 +3,15 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+[0.1.0] - 2023-XX-XX
+--------------------
+* Enhancements
+  * Changed downloads to write files across multiple Instruments, when the
+    remote files contain a mix of data products
+  * Added new instruments: sw_ap, sw_cp, sw_flare, sw_polar-cap, sw_sbfield,
+    sw_ssn, and sw_storm-prob
+  * Added new examples to the documentation
+
 [0.0.9] - 2022-12-21
 --------------------
 * Deprecations